fix: isolate qr scanner search params
